+++ src/layers/l4-controller/commands/index.ts
@@ -14,3 +14,4 @@ export * from './settings';
 export * from './export';
 export * from './resource';
 export * from './syncUpdate';
+export * from './htmlDependency';