+++ src/layers/l4-controller/commands/page/index.ts
@@ -3,5 +3,6 @@ export { UpsertResourceCommand } from './UpsertResourceCommand';
 export type {
   UpsertPageDependencyInput,
   UpsertPageResourceInput,
+  UpsertPageContentInput,
 } from './UpsertResourceCommand';
 export { RecordHtmlDependencyCommand } from './RecordHtmlDependencyCommand';